requirements.txt
Flask
requests
python-dotenv
Pillow
fonttools
pywebview
pyinstaller